[not verified] Licensing: Continued work on moving endpoints from Jetpack plugin to license package
